Refactor getSignedQuote to avoid conditional code
