+++ conclave-enclave/src/main/kotlin/com/r3/conclave/enclave/internal/SocketEnclaveHostInterface.kt
@@ -1,14 +1,14 @@
 package com.r3.conclave.enclave.internal
 
 import com.r3.conclave.common.internal.*
+import com.r3.conclave.common.internal.SgxReport.body
+import com.r3.conclave.common.internal.SgxReportBody.reportData
 import com.r3.conclave.enclave.internal.EnclaveUtils.sanitiseThrowable
 import com.r3.conclave.mail.internal.writeInt
 import com.r3.conclave.utilities.internal.getAllBytes
 import com.r3.conclave.utilities.internal.readIntLengthPrefixBytes
 import com.r3.conclave.utilities.internal.writeIntLengthPrefixBytes
-import java.io.Closeable
-import java.io.DataInputStream
-import java.io.DataOutputStream
+import java.io.*
 import java.net.Socket
 import java.nio.ByteBuffer
 import java.time.Duration
@@ -200,6 +200,7 @@ class SocketEnclaveHostInterface(
      * It is used to link outgoing calls to the appropriate call worker.
      */
     private val threadLocalCallContext = ThreadLocal<EnclaveCallContext>()
+
 
     /**
      * Internal method for initiating a host call with specific arguments.
