+++ conclave-host/src/main/kotlin/com/r3/conclave/host/internal/attestation/EnclaveQuoteServiceMock.kt
@@ -6,8 +6,7 @@ import java.nio.ByteBuffer
 object EnclaveQuoteServiceMock : EnclaveQuoteService() {
 
     override fun initializeQuote(): Cursor<SgxTargetInfo, ByteBuffer> {
-        val targetInfo = Cursor.allocate(SgxTargetInfo)
-        return targetInfo
+        return Cursor.allocate(SgxTargetInfo)
     }
 
     override fun retrieveQuote(report: ByteCursor<SgxReport>): ByteCursor<SgxSignedQuote> {
